refactor(spaces): extract SpacesLogin component from welcome/spaces page
Extract the rendering logic of the /welcome/spaces page into a reusable
SpacesLogin component so it can be shared by the / route in a follow-up.
