+++ apps/web/src/pages/welcome/spaces.tsx
@@ -1,25 +1,6 @@
 import type { NextPage } from 'next'
-import Head from 'next/head'
-import { SpacesFeature, useFeatureFlagRedirect } from '@/features/spaces'
-import { useLoadFeature } from '@/features/__core__'
-import { BRAND_NAME } from '@/config/constants'
-import { useHasFeature } from '@/hooks/useChains'
-import { FEATURES } from '@safe-global/utils/utils/chains'
+import SpacesLogin from '@/features/spaces/components/SpacesLogin'
 
-const Spaces: NextPage = () => {
-  const isSpacesFeatureEnabled = useHasFeature(FEATURES.SPACES)
-  const spaces = useLoadFeature(SpacesFeature)
-  useFeatureFlagRedirect()
-
-  return (
-    <>
-      <Head>
-        <title>{`${BRAND_NAME} – Workspaces`}</title>
-      </Head>
-
-      {isSpacesFeatureEnabled && <spaces.SpacesList />}
-    </>
-  )
-}
+const Spaces: NextPage = () => <SpacesLogin />
 
 export default Spaces
